# web/openapi/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Spades Game Server",
        "description": "4-player Spades card game server with matchmaking, challenges, and real-time updates.",
        "version": "3.0.0"
    },
    "paths": {
...
        "/games/{game_id}/replay.json": {
            "get": {
                "description": "Return a JSON replay of a terminal game. Refused (403) for in-progress\ngames \u2014 the model would expose hidden hands. Resolves `viewer_seat` from\nthe auth `Identity` so clients can orient the replay to the viewer.\n\n`identity` is `OptionalIdentity` (a newtype over `Option<Identity>`) so\nthat a request with a present-but-unrecognized Bearer token (e.g. a\nrevoked API token) still receives the replay with `viewer_seat: null`\nrather than a 401. No-credential and cookie-session paths still resolve\nto `Some(Anonymous)` or `Some(Registered)` as before; only the\ninvalid-Bearer case changes from 401 \u2192 200-with-null-seat.",
                "operationId": "handlers_games_get_replay_json",
                "parameters": [
                    {
                        "name": "game_id",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "in": "path",
                        "style": "simple"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameReplayResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Card": {
                "description": "Intuitive card struct. Ordered by suit (Club < Diamond < Heart < Spade),\nthen rank ascending \u2014 the derived lexicographic order over the fields.",
                "type": "object",
                "properties": {
                    "suit": {
                        "$ref": "#/components/schemas/Suit"
                    },
                    "rank": {
                        "$ref": "#/components/schemas/Rank"
                    }
                },
                "required": [
                    "suit",
                    "rank"
                ]
            },
...
            "DealtHand": {
                "description": "One hand in a `Deal` event: the target seat/pile and the cards dealt to it.",
                "type": "object",
                "properties": {
                    "target": {
                        "type": "string"
                    },
                    "cards": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Card"
                        }
                    }
                },
                "required": [
                    "target",
                    "cards"
                ]
            },
            "Deck": {
                "type": "object",
                "properties": {
                    "suits": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "ranks": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "suits",
                    "ranks"
                ]
            },
...
            "Event": {
                "oneOf": [
                    {
                        "type": "object",
                        "properties": {
                            "hands": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/DealtHand"
                                }
                            },
                            "type": {
                                "type": "string",
                                "enum": [
                                    "deal"
                                ]
                            }
                        },
                        "required": [
                            "hands",
                            "type"
                        ]
                    },
                    {
                        "type": "object",
                        "properties": {
                            "start": {
                                "type": "string"
                            },
                            "values": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "type": {
                                "type": "string",
                                "enum": [
                                    "call"
                                ]
                            }
                        },
                        "required": [
                            "start",
                            "values",
                            "type"
                        ]
                    },
                    {
                        "type": "object",
                        "properties": {
                            "leader": {
                                "type": "string"
                            },
                            "cards": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            },
                            "type": {
                                "type": "string",
                                "enum": [
                                    "play"
                                ]
                            }
                        },
                        "required": [
                            "leader",
                            "cards",
                            "type"
                        ]
                    },
                    {
                        "type": "object",
                        "properties": {
                            "from": {
                                "type": "string"
                            },
                            "to": {
                                "type": "string"
                            },
                            "cards": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            },
                            "type": {
                                "type": "string",
                                "enum": [
                                    "exchange"
                                ]
                            }
                        },
                        "required": [
                            "from",
                            "to",
                            "cards",
                            "type"
                        ]
                    },
                    {
                        "type": "object",
                        "properties": {
                            "target": {
                                "type": "string"
                            },
                            "cards": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            },
                            "type": {
                                "type": "string",
                                "enum": [
                                    "reveal"
                                ]
                            }
                        },
                        "required": [
                            "target",
                            "cards",
                            "type"
                        ]
                    }
                ]
            },
            "GameReplayResponse": {
                "description": "JSON replay response for a terminal game \u2014 available at\n`GET /games/{game_id}/replay.json`.",
                "type": "object",
                "properties": {
                    "model": {
                        "$ref": "#/components/schemas/Model"
                    },
                    "cumulative_by_round": {
                        "type": "array",
                        "items": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "viewer_seat": {
                        "nullable": true,
                        "type": "integer"
                    },
                    "termination": {
                        "$ref": "#/components/schemas/ReplayTermination"
                    }
                },
                "required": [
                    "model",
                    "cumulative_by_round",
                    "termination"
                ]
            },
...
            "Meta": {
                "type": "object",
                "properties": {
                    "version": {
                        "type": "integer"
                    },
                    "game_hint": {
                        "nullable": true,
                        "type": "string"
                    },
                    "seats": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "dealer": {
                        "nullable": true,
                        "type": "string"
                    },
                    "players": {
                        "type": "array",
                        "items": {
                            "nullable": true,
                            "type": "string"
                        }
                    },
                    "partnerships": {
                        "nullable": true,
                        "type": "array",
                        "items": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "caps": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "version",
                    "seats",
                    "players",
                    "caps"
                ]
            },
            "Model": {
                "type": "object",
                "properties": {
                    "meta": {
                        "$ref": "#/components/schemas/Meta"
                    },
                    "deck": {
                        "$ref": "#/components/schemas/Deck"
                    },
                    "events": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Event"
                        }
                    }
                },
                "required": [
                    "meta",
                    "deck",
                    "events"
                ]
            },
...
            "Rank": {
                "type": "string",
                "enum": [
                    "Two",
                    "Three",
                    "Four",
                    "Five",
                    "Six",
                    "Seven",
                    "Eight",
                    "Nine",
                    "Ten",
                    "Jack",
                    "Queen",
                    "King",
                    "Ace"
                ]
            },
            "ReplayTermination": {
                "type": "string",
                "enum": [
                    "completed",
                    "aborted"
                ]
            },
...
            "Suit": {
                "type": "string",
                "enum": [
                    "Club",
                    "Diamond",
                    "Heart",
                    "Spade"
                ]
            },
...
        }
    }
}